+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/io/WebSocketInput.kt
@@ -1,14 +1,14 @@
 package lt.saltyjuice.dragas.chatty.v3.websocket.io
 
+import lt.saltyjuice.dragas.chatty.v3.core.adapter.Deserializer
 import lt.saltyjuice.dragas.chatty.v3.core.io.Input
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Request
 
 /**
  * WebSocket implementation wrapper for core input.
  *
  * @see Input
  */
-interface WebSocketInput : Input<String, Request>
+interface WebSocketInput<Request> : Input<Any, Request>
 {
-
+    override val adapter: Deserializer<Any, Request> get() = throw NotImplementedError("Shouldn't be implmeneted, as this is handled by Decoding layer in tyrus.")
 }