+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/main/WebSocketEndpoint.kt
@@ -9,11 +9,8 @@ import kotlinx.coroutines.experimental.selects.whileSelect
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.AfterMiddleware
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.BeforeMiddleware
 import lt.saltyjuice.dragas.chatty.v3.websocket.adapter.WebSocketAdapter
-import lt.saltyjuice.dragas.chatty.v3.websocket.exception.ServerDestroyedException
 import lt.saltyjuice.dragas.chatty.v3.websocket.io.WebSocketInput
 import lt.saltyjuice.dragas.chatty.v3.websocket.io.WebSocketOutput
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Request
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Response
 import java.io.IOException
 import javax.websocket.*
 
@@ -21,62 +18,75 @@ import javax.websocket.*
 /**
  * Websocket endpoint.
  *
- * This is a programmatically implemented endpoint, which is meant to simplify websocket implementations
+ * This is a programmatically implemented endpoint, which is meant to simplify websocket implementations.
+ *
+ * Implementations should have [ClientEndpoint] annotation, which handles what that endpoint uses as encorder/decoder (adapter)
+ * classes.
  */
-open class WebSocketEndpoint(override val adapter: WebSocketAdapter) : Endpoint(), WebSocketInput, WebSocketOutput
+abstract class WebSocketEndpoint<Request, Response> : Endpoint(), WebSocketInput<Request>, WebSocketOutput<Response>
 {
+
+    final override val adapter: WebSocketAdapter<Any, Request, Response, Any> get() = throw NotImplementedError("Shouldn't be implmeneted, as this is handled by Decoding/Encoding layer in tyrus.")
+    protected open var session: Session? = null
+
     override val beforeMiddlewares: MutableCollection<BeforeMiddleware<Request>> = mutableListOf()
     override val afterMiddlewares: MutableCollection<AfterMiddleware<Response>> = mutableListOf()
 
     override fun getRequest(): Request = runBlocking<Request>
     {
-        return@runBlocking requests.receive()
+        return@runBlocking requests!!.receive()
     }
 
     override fun writeResponse(response: Response)
     {
         launch(CommonPool)
         {
-            responses.send(response)
+            if (afterMiddlewares.firstOrNull { !it.after(response) } == null)
+                responses!!.send(response)
         }
     }
 
-    //protected open val sessions: MutableMap<Session, Job> = Collections.synchronizedMap(HashMap<Session, Job>())
-    init
-    {
-        mDefaultEndpoint = this
-    }
+    /**
+     * Routes requests to [getRequest].
+     */
+    protected open var requests: Channel<Request>? = null
+    /**
+     * Routes responses from [writeResponse] to [responseListener]
+     */
+    protected open var responses: Channel<Response>? = null
+    /**
+     * Listens for responses that come through [writeResponse]
+     */
+    protected open var responseListener: Job? = null
 
-    open val requests: Channel<Request> = Channel(Channel.UNLIMITED)
-    open val responses: Channel<Response> = Channel(Channel.UNLIMITED)
-    protected open val job: Job = launch(CommonPool)
+
+    override fun onOpen(session: Session, config: EndpointConfig)
     {
-        whileSelect()
+        requests = Channel(Channel.UNLIMITED)
+        responses = Channel(Channel.UNLIMITED)
+        this@WebSocketEndpoint.session = session
+        responseListener = launch(CommonPool)
         {
-            responses.onReceive()
+            whileSelect()
             {
-                val rawResponse = adapter.serialize(it)
-                it.session.asyncRemote.sendText(rawResponse)
-                true
+                responses?.onReceive()
+                { response ->
+
+                    session.asyncRemote.sendObject(response)
+                    true
+                }
             }
         }
-    }
-
-    override fun onOpen(session: Session, config: EndpointConfig) = runBlocking<Unit>
-    {
-        session.addMessageHandler(MessageHandler.Whole<String> { message ->
-            val request = adapter.deserialize(message, session)
-            launch(CommonPool)
-            {
-                requests.send(request)
-            }
-        })
     }
 
     @Throws(IOException::class)
     override fun onClose(session: Session, reason: CloseReason)
     {
-        println("Session id ${session.id} closed its connection. Reason: ${reason.closeCode} - ${reason.reasonPhrase}")
+        System.err.println("Session id ${session.id} closed its connection. Reason: ${reason.closeCode} - ${reason.reasonPhrase}")
+        responseListener?.cancel()
+        responseListener = null
+        requests?.close()
+        responses?.close()
     }
 
     @OnError
@@ -85,25 +95,5 @@ open class WebSocketEndpoint(override val adapter: WebSocketAdapter) : Endpoint(
         System.err.println("Something happened for session id ${session.id}")
         throwable.printStackTrace()
         //System.err.println(throwable)
-    }
-
-    /**
-     * This should be called once server's connection is stopped.
-     */
-    open fun onServerDestroyed()
-    {
-        job.cancel(ServerDestroyedException("That's all folks!"))
-    }
-
-    companion object
-    {
-        @JvmStatic
-        private lateinit var mDefaultEndpoint: WebSocketEndpoint
-
-        @JvmStatic
-        fun getDefaultEndpoint(): WebSocketEndpoint
-        {
-            return mDefaultEndpoint
-        }
     }
 }