+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/route/WebSocketMiddleware.kt
@@ -2,13 +2,14 @@ package lt.saltyjuice.dragas.chatty.v3.websocket.route
 
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.AfterMiddleware
 import lt.saltyjuice.dragas.chatty.v3.core.middleware.BeforeMiddleware
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Request
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Response
 
 /**
  * WebSocket wrapper for usual middleware.
+ *
+ * @see BeforeMiddleware
+ * @see AfterMiddleware
  */
-open class WebSocketMiddleware : BeforeMiddleware<Request>, AfterMiddleware<Response>
+open class WebSocketMiddleware<Request, Response> : BeforeMiddleware<Request>, AfterMiddleware<Response>
 {
     override fun before(request: Request): Boolean
     {