+++ websocket/src/main/kotlin/lt/saltyjuice/dragas/chatty/v3/websocket/route/WebSocketRoute.kt
@@ -1,8 +1,6 @@
 package lt.saltyjuice.dragas.chatty.v3.websocket.route
 
 import lt.saltyjuice.dragas.chatty.v3.core.route.Route
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Request
-import lt.saltyjuice.dragas.chatty.v3.websocket.message.Response
 
 
 /**
@@ -10,7 +8,7 @@ import lt.saltyjuice.dragas.chatty.v3.websocket.message.Response
  *
  * @see Route
  */
-abstract class WebSocketRoute : Route<Request, Response>()
+abstract class WebSocketRoute<Request, Response> : Route<Request, Response>()
 {
 
 }